requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
allure-python-commons-test==2.8.16
apipkg==1.5
appdirs==1.4.4
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
distlib==0.3.1
execnet==1.8.0
filelock==3.0.12
idna==2.8
iniconfig==1.1.1
more-itertools==8.7.0
packaging==20.9
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.0.1
pytest-forked==1.3.0
pytest-pycharm==0.7.0
pytest-timeout==1.4.2
pytest-xdist==2.0.0
requests==2.22.0
requests-cache==0.5.2
six==1.15.0
toml==0.10.2
tox==3.14.3
urllib3==1.25.11
virtualenv==20.4.4
